% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_selection_index.R
\name{avg_clust_band_similarity}
\alias{avg_clust_band_similarity}
\title{Average within-cluster band similarity}
\usage{
avg_clust_band_similarity(clust_spec, endpoints_index, collapsed)
}
\arguments{
\item{clust_spec}{A matrix of shape (\code{nfreq}, \code{nrep_clust}) where
\code{nrep_clust} is the number of spectra clustered into the current cluster.}

\item{endpoints_index}{Vector of Fourier indices representing the
frequency bands associated to the current cluster. It is assumed these
include \code{1} and \code{nfreq + 1}.}

\item{collapsed}{Vector of collapsed measures for this cluster.}
}
\value{
Double.
}
\description{
Average the similarity of each band with its right neighbor over all
bands associated to a single cluster.
}
