% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_selection_index.R
\name{band_redundancy}
\alias{band_redundancy}
\title{Band redundancy}
\usage{
band_redundancy(band_spec1, band_spec2, collapsed1, collapsed2, nfreq)
}
\arguments{
\item{band_spec1}{A matrix of shape \verb{(nfreq1, nrep_clust)} where
\code{nfreq1} is the number of frequencies in the first band and \code{nrep_clust}
is the number of replicates in the cluster.}

\item{band_spec2}{A matrix of shape \verb{(nfreq2, nrep_clust)} where
\code{nfreq2} is the number of frequencies in the second band and \code{nrep_clust}
is the number of replicates in the cluster.}

\item{collapsed1}{Double. Collapsed measure of power over the first
band (as estimated by the GA, for example).}

\item{collapsed2}{Double. Collapsed measure of power over the second
band (as estimated by the GA, for example).}

\item{nfreq}{Integer. The number of Fourier frequencies used in estimation.
This is to provide proper scaling for integral approximation.}
}
\value{
Double.
}
\description{
Band redundancy is the ratio of within-band standard deviation of distances
to the same standard deviation computed if the bands were union. Higher
redundancy means that these two quantities are similar in terms of loss
reduction, suggesting that these two bands could be union without significant
loss in approximation quality.
}
