% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_selection_index.R
\name{band_sd}
\alias{band_sd}
\title{(Single) Band standard deviation}
\usage{
band_sd(band_spec, band_collapsed, nfreq)
}
\arguments{
\item{band_spec}{A matrix of shape \verb{(nfreq, nrep_clust)} where
\code{nfreq} is the number of frequencies in the current band and \code{nrep_clust}
is the number of replicates in the cluster.}

\item{band_collapsed}{Double. Collapsed measure of power over the current
band (as estimated by the GA, for example).}

\item{nfreq}{Integer. The number of Fourier frequencies used in estimation.
This is to provide proper scaling for integral approximation.}
}
\value{
Double.
}
\description{
Used in the numerator of band similarity, band standard deviation is defined
as the standard deviation of the distances of each spectra constrained to a
frequency band to the collapsed measure of power computed within that band.
}
