% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_selection_index.R
\name{clust_dist}
\alias{clust_dist}
\title{Distance between two clusters}
\usage{
clust_dist(endpoints_index1, endpoints_index2, collapsed1, collapsed2)
}
\arguments{
\item{endpoints_index1}{Vector of Fourier indices representing the
frequency bands associated to the first cluster being compared. It is
assumed these include \code{1} and \code{nfreq + 1}.}

\item{endpoints_index2}{Vector of Fourier indices representing the
frequency bands associated to the second cluster being compared. It is
assumed these include \code{1} and \code{nfreq + 1}.}

\item{collapsed1}{Vector of collapsed measures for the first cluster being
compared.}

\item{collapsed2}{Vector of collapsed measures for the second cluster being
compared.}
}
\value{
Double.
}
\description{
The distance between two clusters is defined as the L2 distance between
their collapsed spectra.
}
