% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_selection_index.R
\name{clust_sd}
\alias{clust_sd}
\title{Cluster variance}
\usage{
clust_sd(clust_spec, endpoints_index, collapsed)
}
\arguments{
\item{clust_spec}{A matrix of shape (\code{nfreq}, \code{nrep_clust}) where
\code{nrep_clust} is the number of spectra clustered into the current cluster.}

\item{endpoints_index}{Vector of Fourier indices representing the
frequency bands associated to this cluster. It is assumed these include \code{1}
and \code{nfreq + 1}.}

\item{collapsed}{Vector of collapsed measures for this cluster.}
}
\value{
Double.
}
\description{
Cluster variance (used in the numerator of similarity) is defined as the
standard deviation of the L2 distances between each spectra in the cluster
and the cluster-average collapsed spectrum.
}
