% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_selection_index.R
\name{clust_similarity}
\alias{clust_similarity}
\title{Cluster similarity}
\usage{
clust_similarity(
  clust_spec1,
  clust_spec2,
  endpoints_index1,
  endpoints_index2,
  collapsed1,
  collapsed2
)
}
\arguments{
\item{clust_spec1}{A matrix of shape (\code{nfreq}, \code{nrep_clust1}) where
\code{nrep_clust1} is the number of spectra clustered into the current cluster.}

\item{clust_spec2}{A matrix of shape (\code{nfreq}, \code{nrep_clust2}) where
\code{nrep_clust2} is the number of spectra clustered into the current cluster.}

\item{endpoints_index1}{Vector of Fourier indices representing the
frequency bands associated to the first cluster being compared. It is
assumed these include \code{1} and \code{nfreq + 1}.}

\item{endpoints_index2}{Vector of Fourier indices representing the
frequency bands associated to the second cluster being compared. It is
assumed these include \code{1} and \code{nfreq + 1}.}

\item{collapsed1}{Vector of collapsed measures for the first cluster being
compared.}

\item{collapsed2}{Vector of collapsed measures for the second cluster being
compared.}
}
\value{
Double.
}
\description{
Cluster similarity is defined as the ratio of the sum of standard deviations
of the distances of each cluster spectra to their respective cluster centers
(collapsed spectra) to the distance between the collapsed spectra of each
cluster.
}
