% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover}
\usage{
crossover(p1, p2)
}
\arguments{
\item{p1}{Parent 1 - single solution from population}

\item{p2}{Parent 2 - single solution from population}
}
\value{
A list of containing the byproducts of crossover (children) with
names \code{c1} and \code{c2}.
}
\description{
Perform crossover operation between two parents. Crossover is performed by
randomly swapping complimentary subsets of cluster solutions between the
two parents.
}
