% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{dist_mat}
\alias{dist_mat}
\title{Distance matrix}
\usage{
dist_mat(spec, ch)
}
\arguments{
\item{spec}{Matrix of shape (\code{nfreq}, \code{nrep}).}

\item{ch}{A matrix valued solution of shape (\code{nclust}, \code{2 * nbands - 1}).}
}
\value{
Matrix of shape (\code{nrep}, \code{nclust}) where the \verb{(i, j)}th entry is
the L2 distance between the \code{i}th replicate spectrum and the collapsed
spectrum of the \code{j}th cluster.
}
\description{
Return the distance matrix used for cluster assignment by computing the L2
distance between each replicate spectrum and each collapsed spectrum
encoded by \code{ch}.
}
