% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate}
\alias{evaluate}
\title{Population evaluation}
\usage{
evaluate(pop, spec)
}
\arguments{
\item{pop}{A population of solutions as generated by \code{\link{pop_init}}.}

\item{spec}{Matrix of shape (\code{nfreq}, \code{nrep}).}
}
\value{
A vector of length \code{length(pop)} with the fitness values of each
member of the input population.
}
\description{
Evaluate the fitness (inverse of the loss function) of a set of collapsed
cluster-average spectra encoded by one or more solutions contained in a
population.
}
