% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbam.R
\name{fbam}
\alias{fbam}
\title{Frequency band analysis of multiple time series}
\usage{
fbam(
  X,
  nclust_grid,
  nbands_grid,
  parallel = TRUE,
  popsize = 50,
  nparents = 2 * popsize,
  nelite = 2,
  pcrossover = 0.5,
  pmutate = 0.25,
  endpoint_range = NULL,
  collapsed_range = NULL,
  maxgen = 500,
  maxrun = 50,
  tol = 0.001,
  s = NULL,
  ntapers = floor(sqrt(dim(X)[1])),
  verbose = T
)
}
\arguments{
\item{X}{Matrix of shape (\code{n}, \code{nrep}) where \code{n} is the length of the time
series observations and \code{nrep} is the number of independent replicates.}

\item{nclust_grid}{Vector of possible number of clusters. Duplicate values
are removed.}

\item{nbands_grid}{Vector of possible number of bands. Duplicate values are
removed.}

\item{parallel}{Boolean. Should parallelization be used? Defaults to \code{TRUE}.
If a cluster has already been created, that cluster will be used.}

\item{popsize}{Integer. Size of the population carried through each
generation of the GA.}

\item{nparents}{Integer. The number of parents to select that will be
subsequently subjected to crossover and/or mutation in each generation. Must
be at least 2. Defaults to twice the population size.}

\item{nelite}{Integer. The number of elites (most fit) solutions to carry
over from the previous generation into the current generation at each
iteration of the GA. Default is \code{2}. This can be set to \code{0} to prevent
the implementation of an elitism policy.}

\item{pcrossover}{Probability that any two selected parents will be crossed.
Setting this to \code{0} means no crossover will be performed. Default is \code{0.5}.}

\item{pmutate}{Probability that any endpoint or collapsed measure of a single
solution will be subjected to mutation (independent of the other values and
other solutions). Default is \code{0.25}.}

\item{endpoint_range}{Integer or vector of integers. The total maximum width
of the window in which any new endpoint is selected. See details below for
default values.}

\item{collapsed_range}{Double or vector of double. the total width of the
window in which any new collapsed measure is shifted. See details below for
default values.}

\item{maxgen}{Integer. The maximum number of generations the GA is allowed
to run. Defaults to \code{500}.}

\item{maxrun}{Integer. The maxumum number of generations the GA is allowed
to run without significant improvement in the maximum fitness of the
population. Defaults to \code{50}.}

\item{tol}{Double. The percent difference above which a significant
improvement in maximum fitness is declared. Larger values mean that larger
improvements are needed to increase the run count and therefore lead to
shorter convergence times. Defaults to \code{1e-3} (0.1\%).}

\item{s}{the size of the window to shift each evenly spaced frequency within
when initializing endpoints for each solution. Defaults to 10\% of the number
of Fourier frequencies.}

\item{ntapers}{Integer. The number of tapers to use in estimation of the
replicate spectra from \code{X}. Defaults to the floor of the square root of
length of the time series in \code{X}.}

\item{verbose}{Boolean. Whether to print progress of the algorithm to the
console. Defaults to \code{FALSE}.}
}
\value{
A list with the following components.
\item{\code{X}}{Input matrix of time series data}
\item{\code{spec}}{Matrix of shape \verb{(nfreq, nrep)} of spectral estimates used
in optimization.}
\item{\code{labels}}{Clustering labels}
\item{\code{endpoints_index}}{A matrix of shape \verb{(nclust, nbands - 1)}
containing the Fourier indices of the frequency bands associated to each
cluster.}
\item{\code{endpoints}}{A matrix of shape \verb{(nclust, nbands - 1)}
containing the Fourier frequencies of the frequency bands associated to each
cluster.}
\item{\code{endpoints}}{A matrix of shape \verb{(nclust, nbands)}
containing the estimated collapsed measures using the frequency bands
associated to each cluster.}
\item{\code{loss}}{The value of the optimized loss function.}
\item{\code{avgfit}}{A history of the average fitness value in each
generation over the course of evolution.}
\item{\code{maxfit}}{A history of the maximum fitness value in each
generation over the course of evolution.}
\item{\code{ninfeasible}}{A history of the number of infeasible solutions in
each generation over the course of evolution.}
\item{\code{params}}{A list of the input parameters used in the genetic
algorithm(s).}
}
\description{
Frequency band analysis of multiple time series
}
\examples{
data(model1)
# search over a grid of 2 to 4 clusters, and 2 to 6 bands
out <- clustFBE(model1$x, 2:4, 2:6, parallel = T, maxrun = 10)

# fix clusters but vary bands
out <- clustFBE(model1$x, 3, 2:6, parallel = T, maxrun = 10)

# fix both
out <- clustFBE(model1$x, 3, 3, parallel = T, maxrun = 10)

# no clustering
out <- clustFBE(model1$x, 1, 3, parallel = T, maxrun = 10)
}
