% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{fitness}
\alias{fitness}
\title{Fitness evaluation}
\usage{
fitness(ch, spec)
}
\arguments{
\item{ch}{A matrix valued solution of shape (\code{nclust}, \code{2 * nbands - 1}).}

\item{spec}{Matrix of shape (\code{nfreq}, \code{nrep}).}
}
\value{
Double; fitness value (inverse of loss function).
}
\description{
Evaluate the fitness of a single solution (represented by a matrix).
}
