% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm.R
\name{genetic_algorithm}
\alias{genetic_algorithm}
\title{Genetic algorithm}
\usage{
genetic_algorithm(
  spec,
  nbands = 2,
  nsubpop = 1,
  popsize = 50,
  factor = 2,
  pmutate_init = 0.5,
  rate = 50,
  nislands = 6,
  nmigrants = 5,
  epoch = 50,
  maxgen = 500,
  maxrun = 500,
  tol = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{spec}{Replicate-specific spectrum estimates (columns). A single vector is treated as one replicate.}

\item{nbands}{Number of frequency bands.}

\item{nsubpop}{Number of subpopulations.}

\item{popsize}{Number of chromosomes per population.}

\item{factor}{Multiplier of the population size that sets how many offspring to generate each generation for subsequent selection.}

\item{pmutate_init}{Initial mutation rate.}

\item{rate}{Rate at which to decrease mutation rate in terms of number of generations it takes to halve the rate.}

\item{nislands}{Number of independent islands.}

\item{nmigrants}{Number of migrants to move between islands.}

\item{epoch}{How often (in generations) to perform migration.}

\item{maxgen}{Terminate evolution after this number of generations.}

\item{maxrun}{Terminate evolution after non-improvement in fitness after this many generations.}

\item{tol}{Sets the threshold for 'improvement'.}

\item{verbose}{Print information to console?}
}
\value{
A object of class "ga" with the following elements:
\describe{
\item{nfreq}{Number of frequencies (number of rows in 'spec').}
\item{labels}{Subpopulation assignment vector.}
\item{endpoints}{Frequency band endpoints for each subpopulation as integer-valued indices of rows in spec (row-wise).}
\item{avg_summary}{Average summary measures for each subpopulation (rows). Columns correspond to frequency band.}
\item{rep_summary}{Replicate-speciifc summary measures (rows). Columns correspond to frequency bands.}
\item{loss}{Loss function value at solution.}
\item{validation}{Validation criterion evaluated on solution.}
\item{avg_loss}{Vector of average loss function values for each generation.}
\item{min_loss}{Vector of minimum loss function values for each generation.}
\item{ninfeasible}{Number of infeasible solutions in each generation.}
\item{params}{List of algorithm parameters used.}
}
}
\description{
Optimization of FBAM loss function
}
\examples{
x <- matrix(nrow = 128, ncol = 12)
for (i in 1:ncol(x)) x[,i] <- arima.sim(list(ar = runif(1, 0.2, 0.8)), n = 128)
sine_mt_out <- sine_mt(x)
genetic_algorithm(sine_mt_out$mtspec)
}
