% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{loss_function}
\alias{loss_function}
\title{Loss function}
\usage{
loss_function(ch, labels, spec)
}
\arguments{
\item{ch}{A matrix valued solution of shape (\code{nclust}, \code{2 * nbands - 1}).}

\item{labels}{A vector of labels that represent assignments of the spectra
in \code{spec} to clusters. The \code{k}th label represents the row in \code{ch} by which
to collapse the \code{k}th spectrum in \code{spec}.}

\item{spec}{Matrix of shape (\code{nfreq}, \code{nrep}).}
}
\value{
Double; loss function (used to compute fitness).
}
\description{
Evaluate the loss function (sum of within cluster sum of squares) for a
single solution.
}
