% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutate}
\alias{mutate}
\title{Mutation of a single solution}
\usage{
mutate(ch, pmutate, endpoint_range, collapsed_range, nfreq)
}
\arguments{
\item{ch}{A matrix valued solution of shape (\code{nclust}, \code{2 * nbands - 1}).}

\item{pmutate}{Double; probability of mutation of any single Fourier index
or collapsed measure.}

\item{endpoint_range}{Integer; the total maximum width of the window in which
any new endpoint is selected.}

\item{collapsed_range}{Double; the total width of the window in which any
new collapsed measure is shifted.}

\item{nfreq}{Number of frequencies used in estimation. This is used to
ensure new endpoints do not leave the allowed range of Fourier indices.}
}
\value{
A matrix of the same shape as \code{ch} with the new values of
any mutated endpoints or collapsed measures.
}
\description{
Mutation of a single solution proceeds in two stages since each cluster
center encoded by a single solution consists of Fourier indices for the
endpoints and collapsed measures.
}
\details{
Fourier indices (endpoints_index) are each mutated with probability pmutate,
independently of the rest. If an endpoint is selected for mutation, it is
randomly set to a new endpoint that is within a window no wider than
\code{2 * endpoint_range} with the existing endpoint at the center. The
existing endpoints are taken into account so that the new endpoint is not
sampled outside of them resulting in an unordered solution (though this
could probably be relaxed).

Collapsed measures are also each mutated with probability pmutate,
independently of the rest. If a collapsed measure is selected for mutation,
it is randomly shifted by some amount sampled according to a uniform
distribution on \code{-(collapsed_range / 2)} to \code{collapsed_range / 2}. Any
mutated collapsed measures that are mutated to a negative value are
set to zero to ensure validity of the mutated solution.
}
