% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_selection.R
\name{parent_selection}
\alias{parent_selection}
\title{Parent selection}
\usage{
parent_selection(pop, nparents, fitness_values)
}
\arguments{
\item{pop}{A population of solutions as generated by \code{\link{pop_init}}.}

\item{nparents}{Integer. Number of parents from \code{pop} to return. Assumed
that \code{nparents > 1}.}

\item{fitness_values}{Vector of length \code{length(pop)} which gives the
fitness value of each member of the population.}
}
\value{
A list of similar structure to \code{pop} of the selected parents.
}
\description{
Select a specified number of parents from a population using fitness
proportional selection via stochastic universal sampling (SUS).
}
