% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm.R
\name{plot.ga}
\alias{plot.ga}
\title{Print method for object of class "ga"}
\usage{
\method{plot}{ga}(out)
}
\arguments{
\item{out}{An object of class "ga"}
}
\description{
Plot evolution history (average and minimum loss functions in each generation) of the algorithm
}
