% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ga}
\alias{plot_ga}
\alias{plot_ga_solution}
\title{Plot GA evolution history}
\usage{
plot_ga(out)

plot_ga_solution(out)
}
\arguments{
\item{out}{The list object that it outputted by \code{\link{ga}}}
}
\description{
\code{plot_ga()} displays the evolution history in terms of average and minimum
values of the loss function within each generation of a GA run.
\code{plot_ga_solution()} displays a plot of the spectra colored according to
the label they were assigned along with the estimated frequency bands in
each cluster.
}
\examples{
data(model1)
out <- ga(model1$x, 3, 3)
plot_ga(out)
}
