% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_init.R
\name{pop_init}
\alias{pop_init}
\title{Population initialization}
\usage{
pop_init(popsize, nclust, nbands, spec, s = floor(0.1 * dim(spec)[1]))
}
\arguments{
\item{popsize}{Integer. Number of solutions in the population.}

\item{nclust}{Integer. Number of clusters in each solution.}

\item{nbands}{Integer. Number of frequency bands associated to each cluster.}

\item{spec}{Vector of length \code{nfreq} in the case of one replicate or matrix
of shape (\code{nfreq}, \code{nrep}) in the case of multiple replicates where \code{nfreq}
is the number of Fourier frequencies from estimation (typically
\code{floor(T / 2) - 1}).}

\item{s}{the size of the window to shift each evenly spaced frequency within
when initializing endpoints for each solution. Defaults to 10\% of the number
of Fourier frequencies.}
}
\value{
A list of length \code{popsize} where each entry in the list is a matrix
of shape (\code{nclust}, \code{2 * nbands - 1}). Each row of these matrices represents
the cluster center of a single cluster. These rows are organized such that
the first \code{nbands - 1} entries are the indices of the frequency band
boundaries and the remaining entries are the corresponding cluster-average
collapsed measures.
}
\description{
Initialize a population for evolution in a genetic algorithm using
spectral estimates to guide initialization.
}
\details{
Population initialization proceeds in the following way. First, an evenly
spaced grid of \code{nbands + 1} boundaries is set down (this includes the first
index \code{1} and the last index \code{nfreq + 1}). For each solution, these evenly
spaced endpoints are randomly shifted around within a window of width \code{s},
sorted, and used as the initial endpoints for each cluster. To initialize
cluster-average collapsed measures for each cluster, a single replicate is
sampled and collapsed according to these endpoints.

Using the same endpoints for each cluster initially was observed to result
in fewer infeasible solutions, that is, solutions that resulted in empty
clusters when replicates were assigned based on minimum L2 distance.
}
\examples{
# 10 replicates of white noise where T = 400
# spec is simulated using the asymptotic distribution of the periodogram
# for a white noise process with variance 1
length <- 400; nfreq <- floor(length / 2) - 1
nrep <- 10
spec <- matrix(rchisq(nfreq * nrep, df = 2) / 2, nrow = nfreq)
pop <- pop_init(popsize = 100, nclust = 2, nbands = 3, spec = spec)
print(pop[[1]])
}
