% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm.R
\name{print.ga}
\alias{print.ga}
\title{Print method for object of class "ga"}
\usage{
\method{print}{ga}(out, freq = NULL)
}
\arguments{
\item{out}{An object of class "ga"}

\item{freq}{Vector of observed frequencies (useful when sampling rate is different from 1)}
}
\description{
Print method for object of class "ga"
}
