% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsed_by_index.R
\name{rep_collapsed_by_index}
\alias{rep_collapsed_by_index}
\alias{avg_collapsed_by_index}
\title{Collapsed measures from Fourier indices}
\usage{
rep_collapsed_by_index(spec, endpoints_index, labels = NULL)

avg_collapsed_by_index(spec, endpoints_index, labels = NULL)
}
\arguments{
\item{spec}{Vector (one replicate) or matrix of shape (\code{nfreq}, \code{nrep}) where
\code{nfreq} is the number of Fourier frequencies used in estimation
(typically \code{floor(T / 2) - 1}) and \code{nrep} is the number of replicates.}

\item{endpoints_index}{Vector of length \code{nbands - 1} or \code{nbands + 1}
(in the case of one cluster) or matrix (in the case of more than one cluster)
of shape (\code{nclust}, \code{nbands - 1}) or (\code{nclust}, \code{nbands + 1}).
See below for more details.}

\item{labels}{If computing collapsed measures for more than one cluster,
\code{labels} should be a vector of length \code{nrep} where the \code{k}th value points
to the row of \code{endpoints_index} that should be used to collapse the \code{k}th
spectrum (column) in \code{spec}. Default \code{NULL}. If labels are not provided but
endpoints_index includes more than one row, the first row will be used in
collapsing all replicates.}
}
\value{
A matrix of shape (\code{nrep}, \code{nbands}) for replicate-specific collapsed
measures or a matrix of shape (\code{nclust}, \code{nbands}) for cluster-average
collapsed measures.
}
\description{
Computes replicate-specific or cluster-average collapsed measures from
Fourier indices for possibly more than one cluster.
}
\details{
The vector or matrix \code{endpoints_index} can be either of length \code{nbands - 1}
or shape (\code{nclust}, \code{nbands - 1}) or of length \code{nbands + 1} or shape
(\code{nclust}, \code{nbands + 1}). The former case is appropriate when the first and
last indices are NOT included, and the latter case for when they are. The
first index is always \code{1} and the last index is always \code{nfreq + 1}. These are
optional, but if they are included, they must be included for ALL clusters,
that is, in each row of \code{endpoints_index}.
}
\examples{
# 10 replicates of white noise where T = 400
# spec is simulated using the asymptotic distribution of the periodogram
# for a white noise process with variance 1
length <- 400; nfreq <- floor(length / 2) - 1
nrep <- 10
spec <- matrix(rchisq(nfreq * nrep, df = 2) / 2, nrow = nfreq)
labels <- rep(1:2, each = nrep / 2) # equally divided clusters

# first cluster has frequency band endpoints at 50 / T and 100 / T
# second cluster has frequency band endpoints at 70 / T and 120 / T
endpoints_index <- matrix(c(51, 101, 71, 121), nrow = 2, byrow = T)
rep_collapse_by_index(spec, endpoints_index, labels)
}
