% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_index.R
\name{selection_index}
\alias{selection_index}
\title{Cluster/band quantity selection index}
\usage{
selection_index(spec, labels, endpoints_index, collapsed)
}
\arguments{
\item{spec}{A matrix of shape \verb{(nfreq, nrep)} containing all spectra across
all clusters being comapred.}

\item{labels}{Vector of labels.}

\item{endpoints_index}{A matrix of shape \verb{(nclust, nbands - 1)} with the
Fourier indices of the frequency band boundaries for each cluster. It is
assumed these endpoints do NOT contain \code{1} and \code{nfreq + 1} since these are
intended to come directly from the output of the GA which does not carry
these values in its representation.}

\item{collapsed}{A matrix of shape \verb{(nclust, nbands)} with the collapsed
measures of each cluster.}
}
\value{
Double.
}
\description{
Cluster/band quantity selection index
}
