% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine_mt.R
\name{sine_mt}
\alias{sine_mt}
\title{Sine multitaper estimation}
\usage{
sine_mt(X, ntapers = floor(sqrt(dim(as.matrix(X))[1])))
}
\arguments{
\item{X}{Column-wise matrix of replicate time series. A vector is treated as a single replicate.}

\item{ntapers}{Number of tapers to use in constructing the multitaper estimator. Default value is
the floor of the square root of the length of the time series (number of rows in \code{X}).}
}
\value{
A list with the following components:
\itemize{
\item \code{mtfreq}: Vector of frequencies
\item \code{mtspec}: Column-wise matrix of spectral estimates in correspondence to the rows of \code{X}.
}
}
\description{
Compute the multitaper estimator for one or more replicate time series
using the orthogonal sine tapers.
}
\details{
The zero and one-half frequencies are not included in the output estimates.
The choice of \code{ntapers = floor(sqrt(nrow(X)))} ensures consistency of the resulting estimator.
}
\examples{
# single replicate
x <- arima.sim(list(ar = 0.6), n = 500)
sine_mt(x)

# many replicates
X <- matrix(nrow = 500, ncol = 20)
for (i in 1:20) X[,i] <- arima.sim(list(ar = runif(1, 0.2, 0.8)), n = 500)
sine_mt(X)
}
\references{
Riedel, K.S., and A. Sidorenko. 1995. \emph{Minimum Bias Multiple Taper Spectral Estimation.} IEEE Transactions on Signal Processing 43 (1): 188–95.
Available at \url{https://doi.org/10.1109/78.365298}.

Walden, A. T. 2000. \emph{A Unified View of Multitaper Multivariate Spectral Estimation.} Biometrika 87 (4): 767–88.
Available at \url{https://doi.org/10.1093/biomet/87.4.767}.

Percival, Donald B, and Andrew T Walden. 2020. \emph{Spectral Analysis for Univariate Time Series}. Vol. 51. Cambridge University Press.
}
