% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine_multitaper.R
\name{sine_multitaper}
\alias{sine_multitaper}
\title{Multitaper estimation}
\usage{
sine_multitaper(X, ntapers = floor(sqrt(dim(as.matrix(X))[1])))
}
\arguments{
\item{X}{Vector of length \code{n} or matrix of shape (\code{n}, \code{nrep}) where \code{nrep}
is the number of independent time series. If a vector is provided, it is
treated as one series of observations from the same process.}

\item{ntapers}{Integer. The number of tapers to use in estimation. Defaults
to the floor of the square root of the number of observations \code{n}.}
}
\value{
A matrix of shape (\code{floor(n / 2) - 1}, \code{nrep}) even if \code{nrep = 1}.
}
\description{
Multitaper estimation using sine tapers for one or more mutually independent
and stationary time series.
}
\examples{
# single replicate
x <- arima.sim(model(ar = 0.5), n = 1000)
spec <- TSA::ARMAspec(model(ar = 0.5))
mtspec <- sine_multitaper(x)
}
