% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_selection.R
\name{stochastic_universal_sampling}
\alias{stochastic_universal_sampling}
\title{Stochastic Universal Sampling (SUS)}
\usage{
stochastic_universal_sampling(cdf, n)
}
\arguments{
\item{cdf}{cumulative distribution function.}

\item{n}{number of items to select.}
}
\value{
integer indices of selected items.
}
\description{
This particular implementation is based on pseudocode in "Introduction to
Evolutionary Computing", 2e by Eiben and Smith (2015).
}
