% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_selection_index.R
\name{unioned_band_sd}
\alias{unioned_band_sd}
\title{Unioned band standard deviation}
\usage{
unioned_band_sd(union_band_spec, nfreq)
}
\arguments{
\item{union_band_spec}{A matrix of shape \verb{(union_nfreq, nrep_clust)} where
\code{union_nfreq} is the number of frequencies across both bands being compared
and \code{nrep_clust} is the number of replicates in the cluster.}

\item{nfreq}{Integer. The number of Fourier frequencies used in estimation.
This is to provide proper scaling for integral approximation.}
}
\value{
Double.
}
\description{
Used in the denominator of band redundancy, unioned band SD is the
standard deviation of the distances of each spectra constrained to the
union of two neighboring frequency bands to the collapsed measure of
power computed over those two bands.
}
