% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbam.R
\name{fbam}
\alias{fbam}
\title{Frequency band analysis for multiple stationary time series}
\usage{
fbam(x, nbands = 2, nsubpop = 1, ntapers = NULL, ncores = 1, ...)
}
\arguments{
\item{x}{Matrix of replicate time series (columns). A vector is treated as a single replicate.}

\item{nbands}{Integer or vector of numbers of frequency bands to consider.}

\item{nsubpop}{Integer or vector of numbers of subpopulations to consider.}

\item{ntapers}{Number of tapers to use in sine multitaper estimation.}

\item{ncores}{Number of cores to use for parallelization.}

\item{...}{Additional arguments to \link{genetic_algorithm}.}
}
\value{
A list with the following components:
\describe{
\item{grid}{All solutions across the grid of parameters 'nsubpop' > 1 and 'nbands'.}
\item{solution}{Selected solution from 'grid' using validation criteria.}
\item{grid1}{All solutions across the grid of parameters 'nsubpop' = 1 and 'nbands'.}
\item{solution1}{Selection solution from 'grid1' using validation criteria.}
}
}
\description{
Frequency band analysis for multiple stationary time series
}
\examples{
x <- matrix(nrow = 128, ncol = 12)
for (i in 1:ncol(x)) x[,i] <- arima.sim(list(ar = runif(1, 0.2, 0.8)), n = 128)
fbam(x, nbands = 2:5, nsubpop = 1)
fbam(x, nbands = 2:5, nsubpop = 3)
fbam(x, nbands = 2, nsubpop = 1:5)
fbam(x, nbands = 2:5, nsubpop = 1:5)
}
