% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sine_mt.R
\name{sine_mt}
\alias{sine_mt}
\title{Sine multitaper estimation}
\usage{
sine_mt(x, ntapers = NULL)
}
\arguments{
\item{x}{Vector or matrix (columns) of replicate time series. A vector is treated as a single replicate.}

\item{ntapers}{Number of sine tapers. Default is the floor of the square root of the length of the time series (number of rows in \code{x}).}
}
\value{
A list with the following components:
\describe{
\item{mtfreq}{Vector of Fourier frequencies}
\item{mtspec}{Matrix of spectral estimates (columns)}
}
}
\description{
Sine multitaper estimation for one or more replicate time series.
}
\details{
The zero frequency and Nyquist frequency (folding frequency) are not included in the resulting estimates.
}
\examples{
# single replicate
x <- arima.sim(list(ar = 0.6), n = 128)
sine_mt(x)

# many replicates
x <- matrix(nrow = 128, ncol = 8)
for (i in 1:ncol(x)) x[,i] <- arima.sim(list(ar = runif(1, 0.2, 0.8)), n = 128)
sine_mt(x)
}
\references{
Riedel, K.S., and A. Sidorenko. 1995. \emph{Minimum Bias Multiple Taper Spectral Estimation.} IEEE Transactions on Signal Processing 43 (1): 188–95.
Available at \url{https://doi.org/10.1109/78.365298}.

Walden, A. T. 2000. \emph{A Unified View of Multitaper Multivariate Spectral Estimation.} Biometrika 87 (4): 767–88.
Available at \url{https://doi.org/10.1093/biomet/87.4.767}.

Percival, Donald B, and Andrew T Walden. 2020. \emph{Spectral Analysis for Univariate Time Series}. Vol. 51. Cambridge University Press.
}
